Pet model: getImages()
This commit contains the method for retrieving images for a specfic pet.
The code for retriveing the images can be found in `resources->views->pages->pet.blade.php`.
This code only exists to show how the images can be retrieved.

Each pet will have a folder staring with the letter `p` and then its id.
E.g.: `p1`, `p2`, `p3`, and so on.

Images should be named `1`, `2`, `3`, and so on. The file extension doesn't matter.